// src/models/IoK8sApiCoreV1NodeSelectorTerm/types.ts
import { IoK8sApiCoreV1NodeSelectorRequirementModelOptions as IoK8sApiCoreV1NodeSelectorRequirement } from '../IoK8sApiCoreV1NodeSelectorRequirement';

export interface IoK8sApiCoreV1NodeSelectorTermModelOptions {
  matchExpressions?: Array<IoK8sApiCoreV1NodeSelectorRequirement>;
  matchFields?: Array<IoK8sApiCoreV1NodeSelectorRequirement>;
}